separate frontend pages from components
