+++ bike-erp/src/components/OrderService/Components.tsx
@@ -1,29 +1,29 @@
 import axios from "axios";
 import { useState, useEffect } from "react";
 
-import { BACKEND_URL } from "../../../core/utils/config";
+import { BACKEND_URL } from "../../core/utils/config";
 
 import components from "./components.json";
 
-import frame_utility from "../../../assets/images/components/frame_utility.jpg";
-import frame_touring from "../../../assets/images/components/frame_touring.jpg";
-import frame_mountain from "../../../assets/images/components/frame_mountain.jpg";
+import frame_utility from "../../assets/images/components/frame_utility.jpg";
+import frame_touring from "../../assets/images/components/frame_touring.jpg";
+import frame_mountain from "../../assets/images/components/frame_mountain.jpg";
 
-import saddle_performance from "../../../assets/images/components/saddle_performance.jpg";
-import saddle_cushioned from "../../../assets/images/components/saddle_cushioned.jpg";
+import saddle_performance from "../../assets/images/components/saddle_performance.jpg";
+import saddle_cushioned from "../../assets/images/components/saddle_cushioned.jpg";
 
-import handlebar_flat from "../../../assets/images/components/handlebar_flat.jpg";
-import handlebar_bullhorn from "../../../assets/images/components/handlebar_bullhorn.jpg";
-import handlebar_drop from "../../../assets/images/components/handlebar_drop.jpg";
+import handlebar_flat from "../../assets/images/components/handlebar_flat.jpg";
+import handlebar_bullhorn from "../../assets/images/components/handlebar_bullhorn.jpg";
+import handlebar_drop from "../../assets/images/components/handlebar_drop.jpg";
 
-import wheels_utility from "../../../assets/images/components/wheel_utility.jpg";
-import wheels_touring from "../../../assets/images/components/wheel_touring.jpg";
-import wheels_mountain from "../../../assets/images/components/wheel_mountain.jpg";
+import wheels_utility from "../../assets/images/components/wheel_utility.jpg";
+import wheels_touring from "../../assets/images/components/wheel_touring.jpg";
+import wheels_mountain from "../../assets/images/components/wheel_mountain.jpg";
 
-import drivetrain_novice from "../../../assets/images/components/drivetrain_novice.jpg";
-import drivetrain_intermediate from "../../../assets/images/components/drivetrain_intermediate.jpg";
-import drivetrain_advanced from "../../../assets/images/components/drivetrain_advanced.jpg";
-import drivetrain_expert from "../../../assets/images/components/drivetrain_expert.jpg";
+import drivetrain_novice from "../../assets/images/components/drivetrain_novice.jpg";
+import drivetrain_intermediate from "../../assets/images/components/drivetrain_intermediate.jpg";
+import drivetrain_advanced from "../../assets/images/components/drivetrain_advanced.jpg";
+import drivetrain_expert from "../../assets/images/components/drivetrain_expert.jpg";
 
 import {
   Button,
@@ -35,7 +35,7 @@ import {
 } from "@material-ui/core";
 import { grey } from "@material-ui/core/colors";
 import "./Components.css";
-import { addItem, Order } from "../../../redux/actions/OrderListActions/orderListAction";
+import { addItem, Order } from "../../redux/actions/OrderListActions/orderListAction";
 import { connect } from "react-redux";
 
 const WhiteButton = withStyles((theme: Theme) => ({
