+++ bike-erp/src/pages/OrderService/OrderBike/OrderBike.tsx
@@ -1,34 +1,35 @@
 import axios from "axios";
 import { useState, useEffect } from "react";
 
-import { BACKEND_URL } from "../../core/utils/config";
+import { BACKEND_URL } from "../../../core/utils/config";
 
 import components from "./components.json";
 
-import bike_logo from "../../assets/images/login_bike_logo.png";
+import bike_logo from "../../../assets/images/login_bike_logo.png";
 
-import frame_utility from "../../assets/images/components/frame_utility.jpg";
-import frame_touring from "../../assets/images/components/frame_touring.jpg";
-import frame_mountain from "../../assets/images/components/frame_mountain.jpg";
+import frame_utility from "../../../assets/images/components/frame_utility.jpg";
+import frame_touring from "../../../assets/images/components/frame_touring.jpg";
+import frame_mountain from "../../../assets/images/components/frame_mountain.jpg";
 
-import saddle_performance from "../../assets/images/components/saddle_performance.jpg";
-import saddle_cushioned from "../../assets/images/components/saddle_cushioned.jpg";
+import saddle_performance from "../../../assets/images/components/saddle_performance.jpg";
+import saddle_cushioned from "../../../assets/images/components/saddle_cushioned.jpg";
 
-import handlebar_flat from "../../assets/images/components/handlebar_flat.jpg";
-import handlebar_bullhorn from "../../assets/images/components/handlebar_bullhorn.jpg";
-import handlebar_drop from "../../assets/images/components/handlebar_drop.jpg";
+import handlebar_flat from "../../../assets/images/components/handlebar_flat.jpg";
+import handlebar_bullhorn from "../../../assets/images/components/handlebar_bullhorn.jpg";
+import handlebar_drop from "../../../assets/images/components/handlebar_drop.jpg";
 
-import wheels_utility from "../../assets/images/components/wheel_utility.jpg";
-import wheels_touring from "../../assets/images/components/wheel_touring.jpg";
-import wheels_mountain from "../../assets/images/components/wheel_mountain.jpg";
+import wheels_utility from "../../../assets/images/components/wheel_utility.jpg";
+import wheels_touring from "../../../assets/images/components/wheel_touring.jpg";
+import wheels_mountain from "../../../assets/images/components/wheel_mountain.jpg";
 
-import drivetrain_novice from "../../assets/images/components/drivetrain_novice.jpg";
-import drivetrain_intermediate from "../../assets/images/components/drivetrain_intermediate.jpg";
-import drivetrain_advanced from "../../assets/images/components/drivetrain_advanced.jpg";
-import drivetrain_expert from "../../assets/images/components/drivetrain_expert.jpg";
+import drivetrain_novice from "../../../assets/images/components/drivetrain_novice.jpg";
+import drivetrain_intermediate from "../../../assets/images/components/drivetrain_intermediate.jpg";
+import drivetrain_advanced from "../../../assets/images/components/drivetrain_advanced.jpg";
+import drivetrain_expert from "../../../assets/images/components/drivetrain_expert.jpg";
 
 import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline';
 import CheckIcon from '@material-ui/icons/Check';
+
 /*
     This is the bike order page. Users can buy/build different bikes based on what they have in the inventory.
     All bikes must be build with components from the same location and the same size.
@@ -60,7 +61,7 @@ import {
 import { grey } from "@material-ui/core/colors";
 import { ToggleButton, Alert, AlertTitle } from '@material-ui/lab';
 
-import { addBike, BikeSold, removeBike, removeAllBikes, addComponentSold, removeComponentSold, removeAllComponents, ComponentUpdated } from "../../redux/actions/OrderBikeActions/orderBikeActions";
+import { addBike, BikeSold, removeBike, removeAllBikes, addComponentSold, removeComponentSold, removeAllComponents, ComponentUpdated } from "../../../redux/actions/OrderBikeActions/orderBikeActions";
 import { connect } from "react-redux";
 import React from "react";
 import useStyles from "./OrderBikeStyle";
