+++ bike-erp/src/pages/PaymentHistory/ReceivableHistory.css
