+++ bike-erp/src/pages/Registration/RegistrationStyle.ts
