+++ bike-erp/src/test/Inventory/Inventory.test.tsx
@@ -2,7 +2,7 @@ import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import * as React from "react";
 import renderer from "react-test-renderer";
-import Inventory from "../../components/Inventory/inventory";
+import Inventory from "../../pages/Inventory/inventory";
 
 Enzyme.configure({ adapter: new Adapter() });
