+++ bike-erp/src/test/RegistrationPage/RegistrationPage.test.tsx
@@ -1,7 +1,7 @@
 import { configure, shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import renderer from "react-test-renderer";
-import RegistrationPage from "../../components/RegistrationPage/RegistrationPage";
+import RegistrationPage from "../../pages/Registration/Registration";
 
 // Enable enzyme testing
 configure({ adapter: new Adapter() });
